doing something

+++ Items.py
@@ -17,6 +17,7 @@ class DilutedHealingElixir:
         self.value = 4
         self.maxStack = 5
         self.healingPower = [3,7]
+        self.tags = ["potion","healthPotion","unidentifiedElixirUsable"]
 
     def get_value(self):
         return self.value
@@ -48,6 +49,7 @@ class ImpureHealingElixir(DilutedHealingElixir):
         self.value = 10
         self.maxStack = 5
         self.healingPower = [5,10]
+        self.tags = ["potion","healthPotion","unidentifiedElixirUsable"]
 
 class DistilledHealingElixir(DilutedHealingElixir):
     def __init__(self):
@@ -59,14 +61,17 @@ class DistilledHealingElixir(DilutedHealingElixir):
         self.value = 20
         self.maxStack = 5
         self.healingPower = [15,20]
+        self.tags = ["potion","healthPotion","unidentifiedElixirUsable"]
 
 class MysteryElixir:
     def __init__(self):
         self.name = "{}{}Mystery Elixir{}".format(txt.col.fg.nml.magenta,txt.sty.bold,txt.sty.reset)
         self.description = "An unknown elixir. Who knows what it does?"
         self.inspectMessage = "This flask contains a greyish fluid. You cannot see anything through it."
+        self.destroyOnUse = True
         self.value = randint(3,20)
         self.maxStack = 1
+        self.tags = ["potion","unidentifiedElixirUsable"]
 
     def get_type(self):
         return "potion"
@@ -85,25 +90,37 @@ class MysteryElixir:
             print("Too late you realise the recklessness of your drinking as poison spills down your throat, taking its effect in seconds.")
             hero.receiveDamage(1000000000) #there's more elegant ways to kill you, yes. there's also more elegant ways to make a game and i don't see them used here
             return
-        inventory.destroy_general_item([self,1])
+        if self.destroyOnUse:
+            inventory.destroy_general_item([self,1])
 
 class UnidentifiedElixir:
-    def __init__(self):
-        self.potions = [DilutedHealingElixir,ImpureHealingElixir,DistilledHealingElixir,MysteryElixir]
+    def __init__(self,inventory):
+        self.potions = inventory.get_items_with_tags(["unidentifiedElixirUsable"])
         self.identity = self.potions[randint(0,len(self.potions)-1)]
         self.description = "You can't seem to decypher what kind of potion this is..."
+        self.destroyOnUse = True
         self.value = randint(3,20)
         self.maxStack = 1
+        self.tags = ["potion"]
 
     def get_type(self):
         return "potion"
 
-    def use(self,player,inventory):
-        self.identity.use(player,inventory)
+    def use(self,player,inv):
+        potion = self.identity() # virtualises the potion that it is 
+        potion.destroyOnUse = False # makes it so it doesnt destroy an item it doesnt actually have (its virtualised, not in the inventory)
+        potion.use(player,inv) # use the virtualised potion
+        if self.destroyOnUse:
+            inv.destroy_general_item([self,1]) # destroy the unidentified potion, not the virtualised one
 
-    def inspect(self):
+    def generate_inspect_message(self):
+        if randint(1,100) < 30:
+            # 30% chance of not being able to see anything
+            inspectMessage = "You can't make out much through the murky glass of the flask."
+        else:
+            # 70% chance of being able to see something
 
-        if randint(1,100) < 70:
+            # define the different messages for the different potions
             if self.identity == DilutedHealingElixir:
                 messages = ["doesn't smell of much","has a slight tinge of red to it"]
 
@@ -116,14 +133,18 @@ class UnidentifiedElixir:
             elif self.identity == MysteryElixir:
                 messages = ["smells bad","looks almost grey"]
 
+            # invent the english language
             inspectMessage = "The potion "
             conjoiners = [". It ",", it "]
             enders = [".", "..."]
             thoughEnders = [" though.", " though..."]
             usedStatements = []
 
+            # create the inspect message
             length = randint(1,2)
             for i in range(0,length,1):
+
+                # find a message to use that hasnt been used already
                 candidateFound = False
                 while not candidateFound:
                     if len(usedStatements) >= len(messages):
@@ -131,18 +152,31 @@ class UnidentifiedElixir:
                     candidate = randint(0,len(messages)-1)
                     if not candidate in usedStatements:
                         candidateFound = True
+
+                # if we successfully found one:
                 if candidateFound:
                     if i != length and i != 0:
+                        # if its not the end of the message, add a conjunction (conjoiner)
                         inspectMessage += conjoiners[randint(0,len(conjoiners)-1)]
                     inspectMessage += messages[candidate]
                     usedStatements.append(candidate)
+
+            # add the ending of the sentence (the ender)
             if length > 1 and randint(1,100) < 50:
+                # if the message has two points in it, 50% chance of adding "though" to the end
                 inspectMessage += thoughEnders[randint(0,len(thoughEnders)-1)]
             else:
                 inspectMessage += enders[randint(0,len(enders)-1)]
-        else:
-            inspectMessage = "You can't make out much through the murky glass of the flask."
         return inspectMessage
+
+    def inspect(self,inventory):
+        
+        # STILL NEED TO ADD THE CODE FOR CHECKING IF THE POTION'S TRUE IDENTITY IS ALREADY IN THE PLAYER'S INVENTORY!
+        if inventory.is_in_inv(self.identity):
+            if self.destroyOnUse:
+                inventory.destroy_general_item
+        
+        return self.generate_inspect_message()
 
 
 
@@ -157,6 +191,7 @@ class PotionPouch:
         self.inspectMessage = "Use this item to put on the pouch, giving you 5 potion slots."
         self.value = 5
         self.maxStack = 1
+        self.tags = ["clothierSellable"]
 
     def use(self,hero,inv):
         print("You put on the potion pouch, now you can store 5 more types of potion.")
